Syntaxerror/requirements.txt
pygame
pathfinding
pyserial
